[core] Move heap-allocating helpers to alloc_helpers.h/cpp
Move all heap-allocating helper functions from helpers.h/cpp to
dedicated alloc_helpers.h/cpp files. This makes it immediately obvious
when code depends on allocating helpers, as it must include
alloc_helpers.h.

For backward compatibility, helpers.h re-exports alloc_helpers.h.
This re-export will be removed in 2026.11.0. External components
should update their includes before then.

Functions moved:
- str_truncate, str_upper_case, str_snake_case, str_sanitize
- str_snprintf, str_sprintf
- format_mac_address_pretty, format_hex, format_hex_pretty, format_bin
- value_accuracy_to_string
- base64_encode (both overloads), base64_decode (vector overload)
- get_mac_address, get_mac_address_pretty

Also adds base64_encode and value_accuracy_to_string to the CI lint
check for heap-allocating helpers.
